Local Time and Date by using JavaScript
This is my very first project in JavaScript. I did it as a hobby project and for the experiment that, what I have learned up to this. You can see the local date and time here. I will try to implement this as a world clock and date as well.